feat(app): added deployment script for production
This commit makes several changes to the docker setup to create a deployment process for production.  The docker compose files are now split into four files:
	- docker-compose.yml: defines the api and db services
	- docker-compose.dev.yml: defines development specific docker behaviour
	- docker-compose.prod.yml: defines production specific docker behaviour
	- docker-compose.frontend.yml: defines a container used only to build optimized webpack bundles

Environment variables for rails and postgres can be found in prod.env and dev.env.

To deploy, pull the most recent master branch into the server and run the deploy script.  This will spin up the rails and postgres containers in detached mode.
To run the app locally, run the start_dev script.  This will spin up a rails, webpack, and postgres server in development mode with hot-reloading enabled.