[frontend] Added applicants view (#399)
Add an "Applicants" view to the admin page.

This PR also adds an "upsert by session" route for applicants. If this route is used and an applicant is upserted, a blank application associated with the specified session will be created. That way, when you query applicants by session, the newly upserted applicant will show up.